add scanner, in order not to break SRP
