+++ src/test/java/com/epam/prejap/teatrees/game/PrinterTest.java
@@ -15,8 +15,6 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Scanner;
 
-
-
 public class PrinterTest {
 
     @Test
@@ -45,37 +43,6 @@ public class PrinterTest {
 
     }
 
-    @Test(dataProvider = "provideName")
-    public void nameShouldBeSetInConsole(String input, String output) {
-        //given
-        System.setOut(new PrintStream(new ByteArrayOutputStream()));
-        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(input.getBytes(StandardCharsets.UTF_8));
-        Printer printer = new Printer(System.out);
-        //when
-        String actualName = printer.askForUserName(new Scanner(byteArrayInputStream));
-        //then
-        Assert.assertEquals(actualName, output);
-    }
 
-    @Test(expectedExceptions = NoSuchElementException.class)
-    public void rotateInEndlessLoopIfNameIsLongerThan_3_Letters() {
-        //given
-        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream("LONG_NAME".getBytes(StandardCharsets.UTF_8));
-        Printer printer = new Printer(System.out);
-        //when
-        String actualName = printer.askForUserName(new Scanner(byteArrayInputStream));
-        //then
-        Assert.assertEquals(actualName, "LONG_NAME");
-    }
-
-    @DataProvider
-    public Object[][] provideName() {
-        return new Object[][]{
-                {"nad", "nad"},
-                {"sss", "sss"},
-                {"111", "111"},
-                {"   ", "   "},
-        };
-    }
 
 }
